requirements.txt
requests==2.31.0
pytz==2023.3.post1
pushbullet.py==0.12.0